Added cf and bootstrap variants for EFS in case of AWS outage
The Gogs setup is function running off the EFS volume, but it is not
very performant.  These scripts allow a cluster to be booted and
configured to run off EFS, which could be useful if there was an AWS
outage or some particular issue affecting our existing setup.

These could also be used if EFS performance in handling small files
got better or if we tuned Gogs such that it hits the filesystem less
heavily.